requirements.txt
alpaca-py
pandas
ta
requests
google-generativeai
groq
python-dotenv
